+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 
 ## [Unreleased] - ReleaseDate
+
+## [0.9.4] - 2026-02-03
 ### Added
 - Support for `rkyv` based (de)serialization ([#44]).
 
@@ -181,7 +183,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/loyd/fixnum/compare/v0.9.3...HEAD
+[Unreleased]: https://github.com/loyd/fixnum/compare/v0.9.4...HEAD
+[0.9.4]: https://github.com/loyd/fixnum/compare/v0.9.3...v0.9.4
 [0.9.3]: https://github.com/loyd/fixnum/compare/v0.9.2...v0.9.3
 [0.9.2]: https://github.com/loyd/fixnum/compare/v0.9.1...v0.9.2
 [0.9.1]: https://github.com/loyd/fixnum/compare/v0.9.0...v0.9.1